add possibility to chose columns to select when we do a SELECT method

+++ src/main/java/org/dant/model/Condition.java
@@ -51,7 +51,7 @@ public class Condition {
                 if ( type.equals("BINARY") )
                     return ((String)this.value).compareTo((String) list.get(index)) < 0;
                 if ( type.equals("INT64") )
-                    return ((BigDecimal)this.value).longValue() < (Long)list.get(index);
+                    return ((BigDecimal)this.value).byteValue() < (byte)list.get(index);
                 if ( type.equals("DOUBLE") )
                     return ((BigDecimal)this.value).doubleValue() < (Double)list.get(index);
                 break;
@@ -59,7 +59,7 @@ public class Condition {
                 if ( type.equals("BINARY") )
                     return (this.value).equals(list.get(index));
                 if ( type.equals("INT64") )
-                    return ((BigDecimal)this.value).longValue() == (Long)list.get(index);
+                    return ((BigDecimal)this.value).byteValue() == (byte) list.get(index);
                 if ( type.equals("DOUBLE") )
                     return ((BigDecimal)this.value).doubleValue() == (Double)list.get(index);
                 break;
@@ -67,7 +67,7 @@ public class Condition {
                 if ( type.equals("BINARY") )
                     return ((String)this.value).compareTo((String) list.get(index)) > 0;
                 if ( type.equals("INT64") )
-                    return ((BigDecimal)this.value).longValue() > (Long)list.get(index);
+                    return ((BigDecimal)this.value).byteValue() > (byte)list.get(index);
                 if ( type.equals("DOUBLE") )
                     return ((BigDecimal)this.value).doubleValue() > (Double)list.get(index);
                 break;
